fix bugs with frontend filex
