--- client/src/events/components/carousel.js
@@ -1,44 +1,0 @@
-import React, {Component} from 'react';
-
-
-class Carousel extends Component{
-    render(){ 
-        return( 
-            <div id="carouselExampleCaptions" className="carousel slide" data-ride="carousel">
-            <ol className="carousel-indicators">
-              <li data-target="#carouselExampleCaptions" data-slide-to="0" className="active"></li>
-              <li data-target="#carouselExampleCaptions" data-slide-to="1"></li>
-              <li data-target="#carouselExampleCaptions" data-slide-to="2"></li>
-            </ol>
-            <div className="carousel-inner">
-              <div className="carousel-item active">
-                <img src="https://www.acenet.edu/PublishingImages/Interior-Page-Heroes/2018ACE-1045.JPG?RenditionID=10" className="d-block w-100" alt="..."/>
-                <div className="carousel-caption d-none d-md-block">
-                  <h5>Офіційні події</h5>
-                  <p>Заворожуюче прекрасний опис переваг нашого сайту</p>
-                </div>
-              </div>
-              <div className="carousel-item">
-              <img src="https://theoxfordmagazine.com/wp-content/uploads/oxford-festival-of-arts-2019-1920x756.jpg" className="d-block w-100" alt="..."/>
-                <div className="carousel-caption d-none d-md-block">
-                  <h5>Неофіційні події</h5>
-                  <p>неясно пікнік це чи весілля але гарно</p>
-                </div>
-              </div>
-            </div>
-            <a className="carousel-control-prev" href="#carouselExampleCaptions" role="button" data-slide="prev">
-              <span className="carousel-control-prev-icon" aria-hidden="true"></span>
-              <span className="sr-only">Previous</span>
-            </a>
-            <a className="carousel-control-next" href="#carouselExampleCaptions" role="button" data-slide="next">
-              <span className="carousel-control-next-icon" aria-hidden="true"></span>
-              <span className="sr-only">Next</span>
-            </a>
-          </div>
-                
-        )}
-
-};
-
-
-export default Carousel;
